join classroom

+++ domain/src/main/java/com/eng/asu/adaptivelearning/domain/CourseService.java
@@ -13,5 +13,5 @@ public interface CourseService {
 
     Observable<List<FancyCourse>> getCoursesByCategory(String category);
 
-    Observable<FancyCourse> getCourse(Integer courseId);
+    Observable<FancyCourse> getCourse(Long courseId);
 }
